requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
rich==13.7.1
typer==0.12.3
google-genai==0.6.0
python-dotenv==1.0.1
flask==3.0.2
gunicorn==21.2.0
google-generativeai
cloudscraper==1.2.71
